Take 'Archived' prop into account in entities remove methods

+++ FacilityServices/FacilityServices.DataLayer/Repositories/ComponentTypeRepository.cs
@@ -58,31 +58,29 @@ namespace FacilityServices.DataLayer.Repositories
 
         public bool Remove(int Id)
         {
-            if (!facilityContext.ComponentTypes.Any(x => x.Id == Id))
+            if (!facilityContext.ComponentTypes.Any(x => x.Id == Id && x.Archived != true))
                 throw new Exception($"ComponentTypeRepository. Delete(ComponentTypeId = {Id}) no record to delete.");
 
-            var ReturnValue = false;
-
-            var componentType = facilityContext.ComponentTypes.FirstOrDefault(x => x.Id == Id);
+            var componentType = facilityContext.ComponentTypes.FirstOrDefault(x => x.Id == Id && x.Archived != true);
             if (componentType != default)
             {
                 try
                 {
-                    facilityContext.ComponentTypes.Remove(componentType);
-                    ReturnValue = true;
+                    componentType.Archived = true;
+                    return facilityContext.ComponentTypes.Update(componentType).Entity.Archived;
                 }
                 catch (Exception)
                 {
-                    ReturnValue = false;
+                    return false;
                 }
             }
 
-            return ReturnValue;
+            return false;
         }
 
         public ComponentTypeTO Update(ComponentTypeTO Entity)
         {
-            if (!facilityContext.ComponentTypes.Any(x => x.Id == Entity.Id))
+            if (!facilityContext.ComponentTypes.Any(x => x.Id == Entity.Id && x.Archived != true))
                 throw new Exception($"ComponentTypeRepository. Update(ComponentTypeTransfertObject) no record to update.");
 
             var attachedComponentTypes = facilityContext.ComponentTypes
